fixing wrong background color on table cell hover
